backend/requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
werkzeug==2.3.7
pyyaml==6.0.1
flasgger==0.9.7.1
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.3
gunicorn==21.2.0